+++ core/src/commonMain/kotlin/com/powersync/db/internal/InternalDatabase.kt
@@ -6,5 +6,7 @@ import kotlinx.coroutines.flow.SharedFlow
 internal interface InternalDatabase : Queries {
     fun updatesOnTables(): SharedFlow<Set<String>>
 
+    suspend fun updateSchema(schemaJson: String): Unit
+
     suspend fun close(): Unit
 }